requirements.txt
Flask
Flask-login
Flask-SQLAlchemy
flask-bcrypt
pycrypto
gunicorn
mysql-python